Refactored `ReputationFactor` To Use `Double` Instead Of `Int`
- Changed `reputationFactor` data type from `int` to `double` in multiple classes for improved precision.
- Updated `ContractTerms` constructor and `addUnitReputationModifiers` to work with `double` values, including proper flooring and clamping.
- Modified `calculatePaymentMultiplier` in `CamOpsContractMarket` to utilize the new `double` type for `reputationFactor`.
- Adjusted `getReputationFactor` in `ReputationController` to return a `double` for consistency with related changes.

+++ MekHQ/src/mekhq/campaign/market/contractMarket/CamOpsContractMarket.java
@@ -106,7 +106,7 @@ public class CamOpsContractMarket extends AbstractContractMarket {
 
     @Override
     public double calculatePaymentMultiplier(Campaign campaign, AtBContract contract) {
-        int reputationFactor = campaign.getReputation().getReputationFactor();
+        double reputationFactor = campaign.getReputation().getReputationFactor();
         ContractTerms terms = getContractTerms(campaign, contract);
         return terms.getEmploymentMultiplier() * terms.getOperationsTempoMultiplier() * reputationFactor;
     }
